Move Elastica adapter to package
