+++ lib/Adapter/Elastica/composer.json
@@ -1,0 +1,23 @@
+{
+    "name": "pagerfanta/elastica-adapter",
+    "type": "library",
+    "description": "Pagerfanta adapter for Elastica",
+    "keywords": ["pagerfanta", "elastica"],
+    "license": "MIT",
+    "require": {
+        "php": "^7.2",
+        "pagerfanta/core": "^2.4",
+        "ruflin/elastica": "^1.3 || ^2.0 || ^3.0 || ^5.0 || ^6.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "Pagerfanta\\Elastica\\": "./"
+        }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-2.x": "2.x-dev"
+        }
+    },
+    "minimum-stability": "dev"
+}
